% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_utils.R
\name{adstock_geometric}
\alias{adstock_geometric}
\title{Geometric Adstock}
\usage{
adstock_geometric(input, decay)
}
\arguments{
\item{input}{The input vector to be transformed}

\item{decay}{The decay rate parameter. Also called "theta" in other contexts.}
}
\value{
The transformed vector
}
\description{
Applies geometric adstocking to vector
}
\examples{
set.seed(1234)
input <- rnorm(20, 10, 5)

input

adstock_geometric(input, .5)

}
