% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_utils.R
\name{apply_media_transforms}
\alias{apply_media_transforms}
\title{Apply Media Transforms}
\usage{
apply_media_transforms(data, hyps, compute_gammatrans = TRUE)
}
\arguments{
\item{data}{The dataframe to be computed on.}

\item{hyps}{A dataframe storing the hyperparameters. It needs four columns: A "predictors" column identifying the variables to be transformed, an "alphas" column with the alpha parameter for each predictor, a "gammas" columns with the gamma parameter for each predictor, and a "thetas" column for each theta paramter for each predictor. For non-saturated variables, leave alphas and gammas as "NA". For non-adstocked variables leave thetas as NA.}

\item{compute_gammatrans}{Defaults to TRUE. If TRUE, the gammaTrans value for each parameter will be computed for each saturated variable. If false, gammaTrans must be a column in the "hyps" dataframe.}
}
\value{
The gammatrans parameter (a single numeric value)
}
\description{
Given a dataframe, and a dataframe of alpha, gamma, and theta parameters, this function applies adstocking and diminishing returns effects to the original dataframe.
}
