% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabinoffmmmr_class.R
\name{coef.mmmr_fit}
\alias{coef.mmmr_fit}
\title{coef.mmmr_fit}
\usage{
\method{coef}{mmmr_fit}(object, complete = TRUE, params = FALSE, ...)
}
\arguments{
\item{object}{An mmmr object}

\item{complete}{If TRUE, NA coefficients will be included}

\item{params}{If TRUE, a dataframe including the alpha, gamma, and theta parameters will be returned.}

\item{...}{Not currently used}
}
\value{
An S3 object of type mmmr
}
\description{
Get coefficients from an mmmr_fit object
}
