% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabinoffmmmr_class.R
\name{fit.mmmr}
\alias{fit.mmmr}
\title{fit.mmmr}
\usage{
\method{fit}{mmmr}(object, data, maxiter = 10, ...)
}
\arguments{
\item{object}{The mmmr object to be used for fitting.}

\item{data}{A dataframe to be used for fitting.}

\item{maxiter}{Maximum number of iterations for the genetic algorithm. Passed to GA::de}

\item{...}{Other argument passed to GA::de. The ... argument to GA::de are passed to the fitness function, which then gets passed to the glmnet::cv.glmnet function.}
}
\value{
An S3 object of type mmmr_fit
}
\description{
Fits an mmmr model to some data. Returns a \code{fit_mmmr} model object that contains the parameters and coefficients for the solution. A \code{fit_mmmr} object may be used with \code{predict} and \code{coef}.
}
