% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_utils.R
\name{gamma_to_gammatrans}
\alias{gamma_to_gammatrans}
\title{Calculate Gamma Trans}
\usage{
gamma_to_gammatrans(input, gamma, decay = NA)
}
\arguments{
\item{input}{The vector on which to calculate gammatrans}

\item{gamma}{The gamma parameter}

\item{decay}{If non-null, then the input is first transformed with \code{\link[=adstock_geometric]{adstock_geometric()}}}
}
\value{
The gammatrans parameter (a single numeric value)
}
\description{
The gamma parameter controls the point of diminishing returns. For consistency, gamma is given as a number between 0 and 1. Gammatrans is gamma scaled up sensibly to the size of the input.
}
