% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generated_holidays}
\alias{generated_holidays}
\title{Generated Holidays}
\format{
\subsection{\code{generated_holidays}}{

A data frame with 4 columns and 46192 rows:
\describe{
\item{ds}{The date of the holiday}
\item{holiday}{The name of the holiday}
\item{country}{The country code of the country in which the holiday is observed}
\item{year}{The year of the holiday}
}
}
}
\source{
\url{https://github.com/facebook/prophet/tree/main/R/data-raw}
}
\usage{
generated_holidays
}
\description{
The holiday data set from the \code{prophet} package.
}
\keyword{datasets}
