% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_ad_spends}
\alias{historical_ad_spends}
\title{Sample historical spend data}
\format{
\subsection{\code{historical_ad_spends}}{

A data frame with 11 columns and 212 rows:
\describe{
\item{Week Starting}{The date of the week (the monday)}
\item{Sales}{The sales of that week}
\item{TV Spend}{How much was spent on TV ads that ran that week}
\item{Radio Spend}{How much was spent on radio ads that ran that week}
\item{Online Video Spend}{How much was spent on online video ads that ran that week}
\item{Social Media Spend}{How much was spent on social media ads that ran that week}
\item{Search ads Spend}{How much was spent on search ads that ran that week}
\item{Direct Mail Spend}{How much was spent on mailed flyers delivered that week}
\item{black friday sale}{1 if that week includes Black Friday, 0 otherwise}
\item{xmas}{1 if that week includes Christmas day, 0 otherwise}
\item{adverse event}{1 if that week includes some known barrier to sales (construction, extreme weather etc.), 0 otherwise}
}
}
}
\usage{
historical_ad_spends
}
\description{
A contrived example of weekly ad spends. There is a column for each advertising channel, the sales of the store that week, and the date.
}
\keyword{datasets}
