% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabinoffmmmr_class.R
\name{mmm_fitness_gen}
\alias{mmm_fitness_gen}
\title{MMM Fitness Function Preparation}
\usage{
mmm_fitness_gen(
  data,
  dep_col,
  date_col,
  saturated,
  adstocked,
  alphas_low = NULL,
  alphas_high = NULL,
  gammas_low = NULL,
  gammas_high = NULL,
  thetas_low = NULL,
  thetas_high = NULL,
  seed = 1234,
  country = "CA"
)
}
\arguments{
\item{data}{The training data for the model}

\item{dep_col}{The name of the dependent variable}

\item{date_col}{The name of the date variable}

\item{saturated}{The names of the columns which should undergo a diminishing returns transformation.}

\item{adstocked}{The names of the columns which should undergo an adstocking transformation.}

\item{alphas_low}{A named vector of values representing the minimum alpha parameter of some saturated columns.}

\item{alphas_high}{A named vector of values representing the maximum alpha parameter of some saturated columns.}

\item{gammas_low}{A named vector of values representing the minimum gamma parameter of some saturated columns.}

\item{gammas_high}{A named vector of values representing the maximum gamma parameter of some saturated columns.}

\item{thetas_low}{A named vector of values representing the minimum theta parameter of some adstocked columns.}

\item{thetas_high}{A named vector of values representing the maximum theta parameter of some saturated columns.}

\item{seed}{The seed for the random elemens of the function}

\item{country}{The country code for the "Prophet" component of the analysis}
}
\value{
A list. The first element is a list of length 2 containing the "lower" and "upper" arguments to the GA::de function. The second element is a function which takes a solution from the GA::de function and casts it back into named hyperparameter vectors. The third and final value is the fitness function.
}
\description{
Used to generate the inputs for the GA::de function used to calculate saturation and adstock for media channels.
}
