% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabinoffmmmr_class.R
\name{mmmr}
\alias{mmmr}
\title{MMM}
\usage{
mmmr(
  predictors,
  saturated,
  adstocked,
  dep_col,
  date_col,
  alphas_low = NULL,
  alphas_high = NULL,
  gammas_low = NULL,
  gammas_high = NULL,
  thetas_low = NULL,
  thetas_high = NULL,
  seed = 1234,
  country = "CA"
)
}
\arguments{
\item{predictors}{A charactor vector containing the names of the predictor columns of the data.}

\item{dep_col}{A string: the name of the dependent variable of the data.}

\item{date_col}{A string: the name of the date variable of the data.}

\item{alphas_low}{A named numeric vector indicating the minimum value for the alpha parameter for some media channels. If omitted, defaults are used.}

\item{alphas_high}{A named numeric vector indicating the maximum value for the alpha parameter for some media channels. If omitted, defaults are used.}

\item{gammas_low}{A named numeric vector indicating the minimum value for the gamma parameter for some media channels. If omitted, defaults are used.}

\item{gammas_high}{A named numeric vector indicating the maximum value for the gamma parameter for some media channels. If omitted, defaults are used.}

\item{thetas_low}{A named numeric vector indicating the minimum value for the theta parameter for some media channels. If omitted, defaults are used.}

\item{thetas_high}{A named numeric vector indicating the maximum value for the theta parameter for some media channels. If omitted, defaults are used.}

\item{seed}{A numeric value for the seed. Required for reproducibility.}

\item{country}{A string denoting a country code for use with prophet. This code allows the model to use holidays in its time-series component. Country codes are in the \code{country} column of \code{prophet::generated_holidays}.}
}
\value{
An S3 object of type mmmr
}
\description{
Creates a media mix model object. Use \code{fit} to train the model.
}
