% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_plots.R
\name{plot_adstocking}
\alias{plot_adstocking}
\title{Plot Adstocking Curve}
\usage{
plot_adstocking(object, channel, xy_only = FALSE, ...)
}
\arguments{
\item{object}{A fitted mmmr model}

\item{channel}{The name of the media channel to plot}

\item{xy_only}{If TRUE, return a dataframe of the x-y points for the plot. FALSE by default.}

\item{...}{Arguments passed to a theme object for the plot}
}
\value{
If xy_only is false, a ggplot2 object is returned. Otherwise, a dataframe with x-y coordinates is returned.
}
\description{
Produces a ggplot2 object of an adstocking curve for a specific media channel.
}
