% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_plots.R
\name{plot_adstocking_facet}
\alias{plot_adstocking_facet}
\title{Plot Adstocking Curve In A Facet Wrap}
\usage{
plot_adstocking_facet(
  object,
  channels = NULL,
  start_values = NULL,
  xy_only = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted mmmr model}

\item{channels}{The name of the media channels to plot}

\item{start_values}{The starting exposure values for the left-most point on the plots.}

\item{xy_only}{If TRUE, return a dataframe of the x-y points for the plot. FALSE by default.}

\item{...}{Arguments passed to the ggplot2::facet_wrap function}
}
\value{
If xy_only is false, a ggplot2 object is returned. Otherwise, a dataframe with x-y coordinates is returned.
}
\description{
Produces a ggplot2 object of an adstocking curve for a specific media channel.
}
