% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_plots.R
\name{plot_diminishing_returns_facet}
\alias{plot_diminishing_returns_facet}
\title{Plot Diminishing Returns Curves In A Facet Wrap}
\usage{
plot_diminishing_returns_facet(
  object,
  channels = NULL,
  rate = FALSE,
  xy_only = FALSE,
  inflection_point = TRUE,
  x_scale = scales::dollar_format(),
  y_scale = scales::dollar_format(),
  nrow = NULL,
  ncol = NULL,
  scale = "free",
  shrink = TRUE,
  drop = TRUE,
  dir = "h",
  strip.position = "top",
  ...
)
}
\arguments{
\item{object}{A fitted mmmr model}

\item{channels}{The names of the channels to include. If NULL, all relevant channels will be included.}

\item{rate}{If TRUE, the y axis will be the return on the next dollar spent instead of the return on the current spend level.}

\item{xy_only}{If TRUE, return a dataframe of the x-y points for the plot. FALSE by default.}

\item{inflection_point}{If TRUE, the plot will be annotated at the inflection point.}

\item{x_scale}{A scale function applied to the x-axis labels and the x value of the inflection point}

\item{y_scale}{A scale function applied to the y-axis labels and the y value of the inflection point}

\item{nrow, ncol, scale, shrink, drop, dir, strip.position}{Arguments passed to ggplot2::facep_wrap}

\item{...}{Arguments passed to the text geom with the inflection point}
}
\value{
If xy_only is false, a ggplot2 object is returned. Otherwise, a dataframe with x-y coordinates is returned.
}
\description{
Produces a ggplot2 object of a diminishing returns curve for a specific media channel.
}
