% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabinoffmmmr_class.R
\name{predict.mmmr_fit}
\alias{predict.mmmr_fit}
\title{predict.mmmr_fit}
\usage{
\method{predict}{mmmr_fit}(object, newdata = NULL, prophetize = FALSE)
}
\arguments{
\item{object}{An mmmr object}

\item{newdata}{A dataframe to get new predictions from.}
}
\value{
An S3 object of type mmmr
}
\description{
Get predictions from an mmmr_fit object
}
