% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabinoffmmmr_class.R
\name{print.mmmr}
\alias{print.mmmr}
\title{print.mmmr}
\usage{
\method{print}{mmmr}(x, ...)
}
\arguments{
\item{x}{An mmmr object}

\item{...}{Not used}
}
\value{
A character string
}
\description{
Print basic stats about an mmmr model
}
