% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabinoffmmmr_class.R
\name{print.mmmr_fit}
\alias{print.mmmr_fit}
\title{print.mmmr_fit}
\usage{
\method{print}{mmmr_fit}(x, ...)
}
\arguments{
\item{x}{An mmmr_fit object}

\item{...}{Not used}
}
\value{
A character string
}
\description{
Print basic stats about a fitted mmmr model
}
