% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_utils.R
\name{prophetize_df}
\alias{prophetize_df}
\title{Prophetize DF}
\usage{
prophetize_df(
  data,
  dep_col,
  date_col,
  predictors = NULL,
  country = "CA",
  prph = NULL,
  daily.seasonality = FALSE,
  weekly.seasonality = FALSE,
  ...
)
}
\arguments{
\item{data}{The dataframe to be added to.}

\item{dep_col}{The name of the dependent variable.}

\item{date_col}{The name of the date variable.}

\item{predictors}{The names of the predictor variables. If NULL, then all variables will be used other than the dep_col.}

\item{country}{The country code to be used to derive the "holidays" column. Valid country codes include any in the "country" column of prophet::generated_holidays.}

\item{prph}{A prophet model object. If NULL, one will be created.}

\item{daily.seasonality}{If TRUE, daily seasonality will be computed. Currently non-functional.}

\item{weekly.seasonality}{If TRUE, weekly seasonality will be computed. Currently non-functional.}

\item{...}{Other arguments passed to prophet::prophet().}
}
\value{
A dataframe similar to \code{data} but supplemented with \code{trend}, \code{yearly}, and \code{holidays} variables.
}
\description{
Take a dataframe with a date column. Use Meta's "prophet" to add a trend, seasonality, and holidays column to the dataframe.
}
