% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabinoffmmmr-package.R
\docType{package}
\name{rabinoffmmmr-package}
\alias{rabinoffmmmr}
\alias{rabinoffmmmr-package}
\title{rabinoffmmmr: Rabinoff MMMR}
\description{
Make Media Mix Models in R.
}
\author{
\strong{Maintainer}: Daniel Rabinoff \email{d.rabinoff@gmail.com}

}
\keyword{internal}
