% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spend_plan}
\alias{spend_plan}
\title{Spend Plan}
\format{
\subsection{\code{spend_plan}}{

A data frame with 11 columns and 212 rows:
\describe{
\item{Week Starting}{The date of the week (the monday)}
\item{TV Spend}{The planned spend for TV ads that ran that week}
\item{Radio Spend}{The planned spend for radio ads that ran that week}
\item{Online Video Spend}{The planned spend for online video ads that ran that week}
\item{Social Media Spend}{The planned spend for social media ads that ran that week}
\item{Search ads Spend}{The planned spend for search ads that ran that week}
\item{Direct Mail Spend}{The planned spend for mailed flyers delivered that week}
\item{black friday sale}{1 if that week includes Black Friday, 0 otherwise}
\item{xmas}{1 if that week includes Christmas day, 0 otherwise}
\item{adverse event}{1 if that week includes some known barrier to sales (construction, extreme weather etc.), 0 otherwise}
}
}
}
\usage{
spend_plan
}
\description{
A contrived example of weekly planned ad spends. There is a column for each advertising channel of the store that week, and the date.
}
\keyword{datasets}
