% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_plots.R
\name{plot_diminishing_returns}
\alias{plot_diminishing_returns}
\title{Plot Diminishing Returns Curve}
\usage{
plot_diminishing_returns(object, channel, rate = FALSE, xy_only = FALSE, ...)
}
\arguments{
\item{object}{A fitted mmmr model}

\item{channel}{The name of the media channel to plot}

\item{rate}{If TRUE, the y axis will be the return on the next dollar spent instead of the return on the current spend level.}

\item{xy_only}{If TRUE, return a dataframe of the x-y points for the plot. FALSE by default.}
}
\value{
If xy_only is false, a ggplot2 object is returned. Otherwise, a dataframe with x-y coordinates is returned.
}
\description{
Produces a ggplot2 object of a diminishing returns curve for a specific media channel.
}
