% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_utils.R
\name{saturation_hill_trans}
\alias{saturation_hill_trans}
\title{Saturation Hill}
\usage{
saturation_hill_trans(input, alpha, gammatrans, na.rm = TRUE)
}
\arguments{
\item{input}{The input vector to be transformed}

\item{alpha}{The alpha parameter for transformation. This controls the shape of the curve.}

\item{gammatrans}{The gammatrans parameter for transformation. This controls the inflection point of the curve.}

\item{na.rm}{Remove NA values?}
}
\value{
The transformed vector
}
\description{
Applies diminishing returns transformation on media exposure vector
}
